+++ src/options/models/TabModel.ts
@@ -2,6 +2,6 @@ import {Component} from "vue";
 
 export class TabModel {
     title: string;
+    iconSrc: string;
     component: Component;
-    icon: Component;
 }