+++ src/options/models/ToastModel.ts
@@ -1,13 +1,11 @@
-import {Component} from "vue";
-
 export class ToastModel {
     message: string;
     cssClass: string;
-    iconComponent: Component;
+    iconSrc: string;
 
-    constructor(message: string, cssClass: string, iconComponent: Component) {
+    constructor(message: string, cssClass: string, iconSrc: string) {
         this.message = message;
         this.cssClass = cssClass;
-        this.iconComponent = iconComponent;
+        this.iconSrc = iconSrc;
     }
 }