+++ src/shared/enums/IconType.ts
@@ -1,5 +1,4 @@
 export enum IconType {
-    Info = "images/icon-purple-48.png",
     Success = "images/icon-green-48.png",
     Error = "images/icon-red-48.png"
 }