+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+#### v0.0.11
+Extracted SVG images into own files<br/>
+Removed html-webpack-plugin<br/>
+Minor CSS fixes<br/>
+Updated dependencies<br/>
 #### v0.0.10
 Bypass torrent2box if clicked with ctrl, shift or alt keys<br/>
 Replaced Sass with PostCSS<br/>